backend/requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
azure-core==1.31.0
azure-cosmos==4.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
filelock==3.16.1
Flask==2.2.5
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
fsspec==2024.9.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
marshmallow==3.22.0
oauthlib==3.2.2
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dotenv==0.21.1
pytz==2024.1
PyYAML==6.0.2
requests==2.28.2
requests-oauthlib==1.3.1
schedule==1.2.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
tokenizers==0.20.0
tqdm==4.66.5
tweepy==4.14.0
typing_extensions==4.12.2
urllib3==1.26.19
Werkzeug==2.2.3
